+++ tools/preferences/tool.js
@@ -41,7 +41,10 @@
         'backupButtonTitle': 'LiChess Tools - backup preferences in a file',
         'restoreButtonText': 'Restore',
         'restoreButtonTitle': 'LiChess Tools - restore preferences from a file',
-        'defaultValueLegend': '*Blue bordered preferences are the ones enabled by default'
+        'defaultValueLegend': '*Blue bordered preferences are the ones enabled by default',
+        'noDirectoryPickerWarning': 'This browser does not support this functionality',
+        'folderButtonTitle': 'Pick a folder',
+        'fileButtonTitle': 'Pick a file'
       },
       'ro-RO': {
         yes: 'Da',
@@ -68,7 +71,10 @@
         'backupButtonTitle': 'LiChess Tools - Descarc\u0103 preferin\u0163ele \u00eentr-un fi\u015Fier',
         'restoreButtonText': 'Restaurare',
         'restoreButtonTitle': 'LiChess Tools - \u00CEncarc\u0103 preferin\u0163ele dintr-un fi\u015Fier',
-        'defaultValueLegend': '*Preferin\u0163ele cu margine alb\u0103strie sunt cele implicite'
+        'defaultValueLegend': '*Preferin\u0163ele cu margine alb\u0103strie sunt cele implicite',
+        'noDirectoryPickerWarning': 'Acest browser nu suport\u0103 aceast\u0103 func\u0163ionalitate',
+        'folderButtonTitle': 'Alege un director de fi\u015fiere',
+        'fileButtonTitle': 'Alege un fi\u015fier'
       }
     }
 
@@ -220,6 +226,25 @@
                 </div></group>`;
             }
               break;
+            case 'folder': {
+              html += `<group>
+                <div class="folder">
+                  <input class="form-control" type="text" name="${pref.name}" readonly/>
+                  <button class="form-control button picker" type="button" name=${pref.name}-picker" title="$trans(folderButtonTitle)">&#x1F4C1;</button>
+                  <button class="form-control button delete" type="button" name=${pref.name}-delete" title="$trans(folderButtonTitle)">&#x2716;</button>
+                </div></group>`;
+            }
+              break;
+            case 'file': {
+              html += `<group>
+                <div class="file">
+                  <input class="form-control" type="text" name="${pref.name}" readonly/>
+                  <button class="form-control button picker" type="button" name=${pref.name}-picker" title="$trans(folderButtonTitle)" 
+                     data-filedescription="${pref.fileDescription||''}" data-fileextension="${pref.fileExtension||''}">&#x1F4C1;</button>
+                  <button class="form-control button delete" type="button" name=${pref.name}-delete" title="$trans(folderButtonTitle)">&#x2716;</button>
+                </div></group>`;
+            }
+              break;
             default:
               throw new Error('Preference type ' + pref.type + ' not supported');
           }
@@ -310,6 +335,106 @@
             lichess.socket.send('msgSend', { "dest": "totalnoob69", "text": text });
           }
         });
+      $('.folder button.picker', container)
+        .on('click', async ev => {
+          ev.preventDefault();
+          const input=$(ev.target).siblings('input[type="text"]');
+          const name = input.attr('name');
+          if (!name) {
+            throw new Error('Could not find input for folder select button');
+          }
+          let handle;
+          let handleName;
+          if (parent.global.showDirectoryPicker) {
+            handle = await parent.global.showDirectoryPicker({
+              id: 'ltf-'+name,
+              mode: 'readwrite',
+              startIn: 'documents'
+            });
+            handleName = handle?.name;
+          } else 
+          if (parent.global.navigator?.storage?.getDirectory) {
+            handle = await parent.global.navigator.storage.getDirectory();
+            handleName = handle ? 'internal' : '';
+          } else {
+            parent.announce(trans.noarg('noDirectoryPickerWarning'));
+            return;
+          }
+          if (!handle) return;
+          await parent.storage.set('lichessTools/LT/'+name+'-folder', handle, { db: true, raw: true });
+          input
+            .val(handleName)
+            .trigger('change');
+        });
+      $('.folder button.delete', container)
+        .on('click', async ev => {
+          ev.preventDefault();
+          const input=$(ev.target).siblings('input[type="text"]');
+          const name = input.attr('name');
+          if (!name) {
+            throw new Error('Could not find input for folder select button');
+          }
+          await parent.storage.set('lichessTools/LT/'+name+'-folder', undefined, { db: true, raw: true });
+          input
+            .val('')
+            .trigger('change');
+        });
+      $('.file button.picker', container)
+        .on('click', async ev => {
+          ev.preventDefault();
+          const input=$(ev.target).siblings('input[type="text"]');
+          const name = input.attr('name');
+          if (!name) {
+            throw new Error('Could not find input for file select button');
+          }
+          let handle;
+          let handleName;
+          const btn = $(ev.currentTarget);
+          if (parent.global.showOpenFilePicker) {
+            const filetype = {};
+            const description = btn.attr('data-filedescription');
+            if (description) filetype.description = description;
+            const extension = btn.attr('data-fileextension');
+            if (extension) filetype.accept = {
+              "application/octet-stream": [ extension ]
+            }
+            handle= await parent.global.showOpenFilePicker({
+              id: 'ltf-'+name,
+              multiple: false,
+              startIn: 'documents',
+              types: [ filetype ]
+            });
+            handle = handle?.[0];
+            handleName = handle?.name;
+          } else {
+            handle = await new Promise((resolve)=>{
+              $('<input type="file" />')
+                .on('change cancel',e=>{
+                  resolve(e.target.files[0]);
+                })
+                .trigger('click');
+            });
+            handleName = handle?.name;
+          }
+          if (!handle) return;
+          await parent.storage.set('lichessTools/LT/'+name+'-file', handle, { db: true, raw: true });
+          input
+            .val(handleName)
+            .trigger('change');
+        });
+      $('.file button.delete', container)
+        .on('click', async ev => {
+          ev.preventDefault();
+          const input=$(ev.target).siblings('input[type="text"]');
+          const name = input.attr('name');
+          if (!name) {
+            throw new Error('Could not find input for file select button');
+          }
+          await parent.storage.set('lichessTools/LT/'+name+'-file', undefined, { db: true, raw: true });
+          input
+            .val('')
+            .trigger('change');
+        });
       $('div.actionButtons #btnReset', container)
         .on('click', async ev => {
           ev.preventDefault();
@@ -349,7 +474,7 @@
         .on('click', async ev => {
           ev.preventDefault();
           const options = await parent.getOptions();
-          const text = parent.global.JSON.stringify(options, null, 2);
+          const text = parent.global.JSON.stringify({...options, userId: parent.getUserId()}, null, 2);
           const blob = new Blob([text], { type: 'application/json' });
           const url = URL.createObjectURL(blob);
           $('<a>')